Extract test for updating orders
